Adicionado hash nas rotas de newsletter

+++ casa-site/app/Http/Controllers/NewsletterController.php
@@ -2,7 +2,9 @@
 
 namespace App\Http\Controllers;
 
+use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Http\Request;
+use Illuminate\Support\Str;
 use App\Newsletter;
 use Validator;
 use App\Http\Requests\NewsletterRequest;
@@ -33,6 +35,7 @@ class NewsletterController extends Controller
         $dados['receber_eventos'] = true;
         $dados['receber_projetos'] = true;
         $dados['receber_noticias'] = true;
+        $dados['token'] = str_replace('/', '', bcrypt(Str::random(8)));
         
         $newsletter = $this->newsletter->create($dados);
 
@@ -44,14 +47,18 @@ class NewsletterController extends Controller
 
 
     // Método responsavel por abrir a pagina de editar um projeto
-    public function editar($id) 
+    public function editar($id, $token) 
     {
         $registro = $this->newsletter->find($id);
+        if($token != $registro->token) {
+            throw new ModelNotFoundException;
+        } 
+
         return view('admin.newsletter.editar', compact('registro'));
     }
 
     // Método responsavel por salvar as informacoes do formulario de edicao no banco de dados
-    public function atualizar(Request $request,$id) 
+    public function atualizar(Request $request, $id, $token) 
     {
         $dados = $request->all();
 
@@ -74,7 +81,13 @@ class NewsletterController extends Controller
             $dados['receber_noticias'] = false;
         }
         
-        $this->newsletter->find($id)->update($dados);
+        $newsletter = $this->newsletter->find($id);
+        
+        if($token != $newsletter->token) {
+            throw new ModelNotFoundException;
+        } 
+        
+        $newsletter->update($dados);
         
         if(!$dados['receber_eventos'] && !$dados['receber_projetos'] && !$dados['receber_noticias']) {
             $this->newsletter->find($id)->delete();
@@ -85,9 +98,15 @@ class NewsletterController extends Controller
     }
 
     // Metodo da acao de apagar uma newsletter
-    public function deletar($id) 
+    public function deletar($id, $token) 
     {
-        $this->newsletter->find($id)->delete();
+        $registro = $this->newsletter->find($id);
+        if($token != $registro->token) {
+            throw new ModelNotFoundException;
+        } 
+
+        $registro->delete();
+
         return redirect()->back()->with('success', 'Feito! Você não receberá mais novos email.');
     }
 }
